requirements.txt
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-babel==0.6.1
django-filter==1.1.0
django-phonenumber-field==1.3.0
djangorestframework==3.7.3
idna==2.6
olefile==0.44
phonenumberslite==8.8.8
Pillow==4.3.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
